feat: dialog confirmation

+++ src/app/courses/courses/courses.component.ts
@@ -12,6 +12,7 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { CoursesListComponent } from '../courses-list/courses-list.component';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { error } from 'console';
+import { ConfirmationDialogComponent } from '../../shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-courses',
@@ -63,17 +64,27 @@ export class CoursesComponent implements OnInit {
   // }
 
   public onRemove(course: Course) {
-    this.coursesService.remove(course._id).subscribe(
-      () => {
-        this.refresh();
-        this.snackBar.open('Curso removido com sucesso!', 'X', {
-          duration: 5000,
-          verticalPosition: 'top',
-          horizontalPosition: 'center'
-        })
-      },
-      () => this.onError('Erro ao remover curso.')
-    )
+
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certeza que deseja remover esse curso?',
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+
+      if(result) {
+        this.coursesService.remove(course._id).subscribe(
+          () => {
+            this.refresh();
+            this.snackBar.open('Curso removido com sucesso!', 'X', {
+              duration: 5000,
+              verticalPosition: 'top',
+              horizontalPosition: 'center'
+            })
+          },
+          () => this.onError('Erro ao remover curso.')
+        )
+      }
+    });
   }
 
   public refresh() {
